requirements.txt
fastapi==0.129.0
uvicorn==0.41.0
geopandas>=1.1.1
numpy>=2.4.2
pandas>=3.0.1
shapely>=2.1.2
pyproj>=3.7.2
scipy==1.17.0
scikit-learn==1.8.0
xgboost==3.2.0
shap==0.50.0
torch==2.10.0
torchvision==0.25.0
rasterio==1.5.0
requests==2.32.5
pillow==12.1.1
python-dotenv==1.2.1
pyyaml==6.0.3
httpx==0.28.1
